* 'renewal' of https://github.com/rlaalsrl715/naegahama:
  요청글 잔여시간 조회
  s3 삭제 및 요청글 상세조회 수정
  답변글 작성 확인
  요청글 작성 및 임시 저장 구현
  renewal
  message
  "meessage modify"
  merge-again
  설문 num에서 result 로 변경
  Answer 답변 글 초기 생성 시 star 값 null에서 0으로 변경
  설문 시 메달 획득 테스트
  업적 1 , 2, 4, 5, 7, 9 구현
  임시 저장
  회원 가입 시 기본 업적값 0 설정
  post like
  banner 구현
  test data
  test data

# Conflicts:
#	.gradle/7.1/executionHistory/executionHistory.lock
#	.gradle/7.1/fileHashes/fileHashes.bin
#	.gradle/7.1/fileHashes/fileHashes.lock
#	src/main/java/com/hanghae/naegahama/service/SurveyService.java

+++ src/main/java/com/hanghae/naegahama/dto/answer/StarPostRequestDto.java
@@ -3,9 +3,10 @@ package com.hanghae.naegahama.dto.answer;
 import lombok.Getter;
 
 @Getter
+
 public class StarPostRequestDto
 {
-    private Long star;
+    private Integer star;
 
 }
